Add variables list to docs
