+++ src/duqtools/list_variables.py
@@ -1,35 +1,9 @@
-from collections import defaultdict
-from typing import Any, Callable, Dict, Hashable, Iterable, List
-
 import click
 from pydantic import ValidationError
 
 from duqtools.config import cfg, var_lookup
 
-
-def groupby(iterable: Iterable,
-            keyfunc: Callable) -> Dict[Hashable, List[Any]]:
-    """Group iterable by key function.
-    The items are grouped by the value that is returned by the `keyfunc`
-    Parameters
-    ----------
-    iterable : list, tuple or iterable
-        List of items to group
-    keyfunc : callable
-        Used to determine the group of each item. These become the keys
-        of the returned dictionary
-
-    Returns
-    -------
-    grouped : dict
-        Returns a dictionary with the grouped values.
-    """
-    grouped = defaultdict(list)
-    for item in iterable:
-        key = keyfunc(item)
-        grouped[key].append(item)
-
-    return grouped
+from .utils import groupby
 
 
 def list_variables(*, config, **kwargs):
